move asn1 parsing code into util module
