+++ packages/client/src/util/asn1/length.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { ByteStream } from '../../util/stream';
+import { ByteStream } from '../stream';
 import { ASN1ParseError } from './error';
 
 // Decodes the length of a DER-encoded ANS.1 element from the supplied stream.
