+++ packages/client/src/x509/ext.ts
@@ -13,8 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+import { ASN1Obj } from '../util/asn1';
 import { ByteStream } from '../util/stream';
-import { ASN1Obj } from './asn1/obj';
 import { SignedCertificateTimestamp } from './sct';
 
 // https://www.rfc-editor.org/rfc/rfc5280#section-4.1
